base/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
google-cloud-aiplatform==1.104.0
google-cloud-storage==2.19.0
google-cloud-pubsub==2.31.0
beautifulsoup4==4.13.4
google-cloud-secret-manager==2.24.0
requests==2.32.4
lxml==6.0.0
python-dotenv==1.1.1
Flask-Cors==6.0.1
google-genai==1.29.0
tenacity==9.1.2